requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
requests>=2.28.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.0
